Continued implementing proper collisions for characters (WIP)
The character will not need a collider at first, just a single collision (for
all their sprites). A collision is yet to be renamed to a solid (makes more
sense as a solid is a collection of collisions).

Introduced `PositionedSolid` which is the body of e.g. a character. The
navigator still needs to be updated for working with that.

This will allow for characters containing complex collisions shapes (and not
just a rectangle).

This no more compile because of `Navigator` not yet handling `PositionedSolid`.