More on implementing proper collisions for characters (WIP)
Started making Navigator aware of solids, it does not yet work as it should
(some confusions between positions and bounding boxes).

The term "bounding box" is used for both the box defining the bounds of all the
collisions inside a solid, and the bounds of those positioned absolutely in the
level. This should probably be improved.