Correction to the ellipse/ellipse algorithm
The bounding boxes containing one another did not mean an ellipse contained the
other. The reasoning was wrong since a small ellipse in the corner of the
bounding box of a large ellipse would not collide.

Replaced this test by a center check. If the center of the ellipse (initially
the first ellipse) is in the circle (initially the second ellipse), then either
the latter contains the former (which is the case that the bounding boxes were
trying to solve) or they intersect (so returning a collision in those cases is
still right, it would have been reported later anyway).