Fixed all the recent FIXMEs in `navigator.cpp`.
Found a bug when walking along the top of the fences: the character stays
blocked on an invisible wall (when continuously pressing left and down for
example), or go up by one pixel on its own (due to side-stepping) when only
pressing left. This seems to have appeared with the ellipses to ellipses
collisions.