Started working on collision between characters
But for now it is broken and suboptimal. There is something really suspicious
with the positioned ellipses and rectangles of a positioned solid whose
positions are not updated when the position of the solid is updated. This could
cause the positions of the shapes to never change, causing the characters to
never collide between them. But in that case, how the collision with tiles
works?

This commit is mostly a draft and could probably be reverted if restarting from
scratch on this subject.