Started replacing sprite indices (`uint16_t`) by a class `Sprite`
However it does not build for now, and it may be more complicated than
anticipated: if the sprite instances are to be unique for a single sprite (which
is a requirement if the tiles are going to hold references to characters later
on), then the levels should own them. However there are plenty of objects
instanciating them before the levels are created, which is a problem.